+++ pytests/test_common.py
@@ -166,3 +166,16 @@ def test_module_import_microgrid_components() -> None:
     from frequenz.api.common.v1.microgrid.components import transformer_pb2_grpc
 
     assert transformer_pb2_grpc is not None
+
+
+def test_module_import_location() -> None:
+    """Test that the modules can be imported."""
+    # pylint: disable=import-outside-toplevel
+    from frequenz.api.common.v1 import location_pb2
+
+    assert location_pb2 is not None
+
+    # pylint: disable=import-outside-toplevel
+    from frequenz.api.common.v1 import location_pb2_grpc
+
+    assert location_pb2_grpc is not None