+++ RELEASE_NOTES.md
@@ -57,6 +57,8 @@
 
 - Added a message `Microgrid` to represent a microgrid.
 
+- Updated the package name of `location.proto` to `frequenz.api.common.v1`.
+
 ## New Features
 
 <!-- Here goes the main new features and examples or instructions on how to use them -->